[BUGFIX] Use a common index across widget areas
The index was renewed for each list. This didn't used to be a problem
as the settings field used to contain the (now non-existant) embed key,
which typically was unique (although if it ever wasn't unique due to
very bad luck then this same bug would be triggered).

Instead, use a common counter for the index into the various sub-lists
and also when deciding the name for the settings field.