Remove auto-detection of web directory
It was sometimes unrelibable and often did silly things
with the src/ directory.

Also, with hosting (incl vps) becoming increasingly cheap,
the requirement to support installs running within a
directory is a lesser requirement than it used to be. This
type of install is still supported; it's just no longer
fully automatic.