Library for TOTP (Google Authenticator) code generation/checking
Original source was at [1] but I've cleaned up the code quite a
bit and also documented it. The FixedBitNotation class in the origin
repository didn't have a license attached but I managed to find
the original [2] which was since renamed.

In both cases the sources appear to be MIT.

I've put the full license text in the header, as well as adding
Karmabunny as a copyright line for the GoogleAuthenticator class.

To clarify - my modifications to the class are licensed MIT.

[1] https://github.com/chregu/GoogleAuthenticator.php
[2] https://github.com/ademarre/binary-to-text-php/commit/d44968b55fada4c1fac0902f7595b23d2814b2f2