Bad password list to be case sensitive
The point of this check is that attackers will often grab a list
of the most common 1,000 or most common 10,000 and use that as a
start prior to a bruteforce search.

I'd expect this initial attack would be case-sensitive, because
it's designed to "try and get in quick" especially in the case
of an online attack where guess speed is slow.

So taking the top-ten-thousand list and expanding it for all
posible case permuations (~250,000 passwords) would defeat the
purpose of attacking using one of these lists anyway.

And if they're doing an *offline* attack then they've got as much
check speed as they want to pay for so they'll skip this step and
go straight for a brute-force search.