Grant CLI unrestricted access to AdminController
The problem here is the cron system uses Sprout::instance to instantiate
AdminController, which hits the AdminAuth::checkLogin line and attempts to
redirect. This obviously doesn't work under CLI and furthermore because
of the error handling in Sprout::instance it gives the appearance of
fatal error i.e.

<b>FATAL ERROR:</b><br>Unable to instance class "Sprout\Controllers\AdminController".

See the comment on line 220 of Sprout.php for the logic behind this
text. A redirect firing in the constructor of the instantiated class
will break (or confuse if you rather) this trick however:

ob_start();
echo '<preemptive error text>';
new $class_name;
<$class_name::__construct>
    ...
    Url::redirect('...')
        ...
        Event::run('system.shutdown')
            Kohana::shutdown()
            Kohana::closeBuffers($flush = true)

Ideally we'd just add cronGenericActivate to $methods_no_auth but
because we're manually calling a controller method Router::$method isn't
correct.