Tighten up how subsites are found, defaulted, and skipped.
It's terribly unpleasant experience if there are no subsites or
something is missing. Admin view MUST be able to skip these checks
so one can FIX the missing subsites.

Something not fixed here is if a session is 'stuck' on a subsite ID
that no longer exists - everything blows up. But I imagine this is
rare however.