Rework sprout config loading.
Because kohana configs are merged together, there are scenarios where
we have cached half of a complete config because we have attempted to
load it too soon.

This removes any short caches for 'sprout' configs in particular as
this config requires DB requests to determine the appropriate subsite
and there's obviously config things to do well before we're for that.

Previously, if anything loaded 'sprout' _before_ subsites were ready:

- load sprout, cached (without subsite)
- subsites loaded
- load sprout again, hits cache (still without subsite)
- skins breaks, chaos

Now, having controlled caching:

- load sprout, not cached (no subsite)
- subsites loaded
- load sprout again (with subsite)
- everything is OK!