Move resizeUrl() to the base files backend.
This isn't (currently) specific to any one backend. We'll leave
it here for implementations to perhaps cache results or modify
the behaviour for whatever reason.

We've also moved the redirect/replacement logic out of the backend
into the files helper, as this is tied too closely to the files
table. Overall this should better support multi-instance backends.